requirements.txt
pillow >= 11.0.0
pip >= 25.0
setuptools >= 75.8.0
